+++ src/components/CreateArea/index.tsx
@@ -203,5 +203,5 @@ export const CreateArea: FC<ICreateArea> = ({
         }
     };
 
-    return <Wrapper apiKey={resolved.apiKey!} render={renderMap} libraries={[...resolved.libraries]} {...(resolved.version ? { version: resolved.version } : {})}{...(resolved.language ? { language: resolved.language } : {})}{...(resolved.region ? { region: resolved.region } : {})} />;
+    return <Wrapper apiKey={resolved.apiKey!} render={renderMap} libraries={Array.from(new Set([...resolved.libraries]))} {...(resolved.version ? { version: resolved.version } : {})}{...(resolved.language ? { language: resolved.language } : {})}{...(resolved.region ? { region: resolved.region } : {})} />;
 };